+++ CHANGELOG.md
@@ -14,6 +14,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `SchedulerExtension`
 	- `jobs > <id> > enabled` option - allows disabling job via configuration
 
+### Fixed
+
+- support for expression aliases like `'@yearly'`
+
 ## [1.1.0](https://github.com/orisai/nette-scheduler/compare/1.0.2...1.1.0) - 2024-01-26
 
 ### Added